requirements.txt
python-dotenv
schedule
requests
telebot
cryptography